chore(text): cleanup include + diagnostic format (code review #3 + #6)
- content/examples/text/text_animations.cpp: add explicit #include <filesystem>
  next to other stdlib includes. The file uses std::filesystem::current_path()
  in shared_typewriter_engine() but previously relied on transitive inclusion
  via <chronon3d/assets/asset_resolver.hpp> — fragile, breaks if the asset
  resolver header is refactored to drop its <filesystem> include.

- content/common/text_reveal_helpers.hpp::layout_glyphs(): cast f32 font_size
  to int in the shaping-failure diagnostic so output is 'font_size=72' instead
  of 'font_size=72.000000'. Both substr(0, 60) call sites (layout_glyphs +
  build_text_reveal_line) gain a one-line comment marking the 60-char
  truncation as intentional (no ellipsis, keeps log lines bounded).

Single atomic commit, behavior unchanged in non-error path. Refs: code-reviewer
round-6 issues #3 (missing include) and #6 (diagnostic format).

+++ content/common/text_reveal_helpers.hpp
@@ -128,7 +128,9 @@ inline std::vector<GlyphPos> layout_glyphs(
         throw std::runtime_error(
             "layout_glyphs: HarfBuzz shaping produced zero glyphs. "
             "font_path='" + spec.font_path +
-            "' font_size=" + std::to_string(font_size) +
+            // font_size is f32; cast to int for compact diagnostics (avoid "72.000000")
+            "' font_size=" + std::to_string(static_cast<int>(font_size)) +
+            // truncate text to 60 chars (intentional, no ellipsis — keeps log lines bounded)
             " text='" + text.substr(0, 60) + "'. "
             "Check that the font file exists and the AssetResolver is mounted.");
     }
@@ -222,6 +224,7 @@ inline void build_text_reveal_line(SceneBuilder& s,
             "build_text_reveal_line: SceneBuilder has no FontEngine. "
             "Ensure the renderer or composition wires a FontEngine "
             "before calling build_text_reveal_line(). "
+            // truncate text to 60 chars (intentional, no ellipsis — keeps log lines bounded)
             "Text: '" + d.text.substr(0, 60) + "'");
     }
     auto chars = layout_glyphs(d.text, d.font_size, d.font_spec,
